+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatButtonUI.java
@@ -67,6 +67,7 @@ import com.formdev.flatlaf.util.UIScale;
  * @uiDefault Button.focusedBackground			Color	optional
  * @uiDefault Button.hoverBackground			Color	optional
  * @uiDefault Button.pressedBackground			Color	optional
+ * @uiDefault Button.disabledBackground			Color	optional
  * @uiDefault Button.disabledText				Color
  * @uiDefault Button.default.background			Color
  * @uiDefault Button.default.startBackground	Color	optional; if set, a gradient paint is used and Button.default.background is ignored
@@ -97,6 +98,7 @@ public class FlatButtonUI
 	protected Color focusedBackground;
 	protected Color hoverBackground;
 	protected Color pressedBackground;
+	protected Color disabledBackground;
 	protected Color disabledText;
 
 	protected Color defaultBackground;
@@ -142,6 +144,7 @@ public class FlatButtonUI
 			focusedBackground = UIManager.getColor( prefix + "focusedBackground" );
 			hoverBackground = UIManager.getColor( prefix + "hoverBackground" );
 			pressedBackground = UIManager.getColor( prefix + "pressedBackground" );
+			disabledBackground = UIManager.getColor( prefix + "disabledBackground" );
 			disabledText = UIManager.getColor( prefix + "disabledText" );
 
 			if( UIManager.getBoolean( "Button.paintShadow" ) ) {
@@ -356,7 +359,7 @@ public class FlatButtonUI
 
 	protected Color getBackground( JComponent c ) {
 		if( !c.isEnabled() )
-			return null;
+			return disabledBackground;
 
 		// toolbar button
 		if( isToolBarButton( c ) ) {